strandsagent/requirements.txt
strands-agents
bedrock-agentcore
boto3
python-dotenv
pandas
scipy
scikit-learn
streamlit
